chore: rename to eq / multilinear / univariate
